feat: Add return values support
